+++ src/RSoft.Framework/Cross/Abstractions/ServiceActivator.cs
@@ -1,0 +1,46 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using System;
+
+namespace RSoft.Framework.Cross.Abstractions
+{
+
+    /// <summary>
+    /// Add static service resolver to use when dependencies injection is not available
+    /// </summary>
+    public class ServiceActivator
+    {
+        
+        #region Local objects/variables
+
+        private static IServiceProvider _serviceProvider = null;
+
+        #endregion
+
+        #region Public methods
+
+        /// <summary>
+        /// Configure ServiceActivator with full serviceProvider
+        /// </summary>
+        /// <param name="serviceProvider">Service DI provider</param>
+        public static void Configure(IServiceProvider serviceProvider)
+        {
+            _serviceProvider = serviceProvider;
+        }
+
+        /// <summary>
+        /// Create a scope where use this ServiceActivator
+        /// </summary>
+        /// <param name="serviceProvider">Service DI provider</param>
+        public static IServiceScope GetScope(IServiceProvider serviceProvider = null)
+        {
+            var provider = serviceProvider ?? _serviceProvider;
+            return provider?
+                .GetRequiredService<IServiceScopeFactory>()
+                .CreateScope();
+        }
+
+        #endregion
+
+    }
+
+}