+++ src/RSoft.Framework/Domain/Contracts/DateValidationContract.cs
@@ -20,7 +20,6 @@ namespace RSoft.Framework.Domain.Contracts
         public DateValidationContract(DateTime? date, string field, string message) : base()
         {
 
-            //BACKLOG: Globalization
             Contract
                 .IsNotNull(date, field, message)
             ;