+++ docs/release-notes.md
@@ -1,5 +1,8 @@
-# RSoft.Auth
+# RSoft.Framework
 Package that brings together the abstractions and common codes for the application of the design-pattern (DDD standard) model adopted by RSoft
 
 ### Release Notes version 1.0.0-rc1.5
 - Upgrade RSoft.Logs to 1.0.0-rc1.7
+
+### Release Notes version 1.0.0-rc1.6
+- Create extensions to Services Language/Culture/StringLocalizer